// lib/types.ts
export interface BaseDoc {
  id: string;
  createdAt?: Timestamp | Date | string;
  updatedAt?: Timestamp | Date | string;
  isVoided?: boolean;
}

export interface AppUser extends BaseDoc {
  uid: string;
  email: string;
  name: string;
  role: "admin" | "user";
  permissions: string[];
}